As our test count increases, we need more structure to organize it. Split controller tests by functionality.
